requirements.txt
streamlit
python-dotenv
pandas
streamlit>=1.28
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0